+++ game/lib/fxcpds/sprite_customizer/components/sprite_ren.py
